+++ Core/Executor/ContentTypeGroupManager.php
@@ -121,7 +121,7 @@ class ContentTypeGroupManager extends RepositoryExecutor implements MigrationGen
      * @throws \InvalidArgumentException When trying to assign a reference to an unsupported attribute
      * @return array key: the reference names, values: the reference values
      */
-    protected function getReferencesValues($object, array $references)
+    protected function getReferencesValues($object, array $references, $lang)
     {
         $refs = array();