+++ Core/Executor/SectionManager.php
@@ -124,7 +124,7 @@ class SectionManager extends RepositoryExecutor implements MigrationGeneratorInt
      * @throws \InvalidArgumentException When trying to assign a reference to an unsupported attribute
      * @return array key: the reference names, values: the reference values
      */
-    protected function getReferencesValues($section, array $references)
+    protected function getReferencesValues($section, array $references, $lang)
     {
         $refs = array();